all python files are fully typed

+++ web/routes/blog.py
@@ -1,4 +1,5 @@
 from flask import session, redirect, url_for, request
+from werkzeug import Response
 from flask_wtf import FlaskForm
 from web.misc import render_template
 from db import Query
@@ -7,7 +8,7 @@ import profile
 import messaging
 
 
-def blog(user_id: int, post_id: int):
+def blog(user_id: int, post_id: int) -> Response | str | tuple:
     blogpost: dict | None = profile.get_blogpost(user_id, post_id)
     if not blogpost:
         return redirect(url_for("blog_list", user_id=user_id))
@@ -22,33 +23,35 @@ def blog(user_id: int, post_id: int):
                 session["user_id"], post_id, comment_form.corpus.data)
 
         elif delete_form.validate_on_submit():
-            delete_value: str = request.form.get("delete")
+            delete_value: str | None = request.form.get("delete")
             query = Query()
 
             if delete_value == "blog":
                 query.delete_blogpost(user_id, post_id)
 
                 return redirect(url_for("user_profile", user_id=user_id))
-
-            res: str = str(query.get_blog_comment_author(
-                user_id,
-                post_id,
-                delete_value
-            ))
-
-            if (session["user_id"] == user_id) or (session["user_id"] == res):
-                query.delete_blog_comment(post_id, delete_value)
-
-                return redirect(url_for(
-                    "blog",
-                    user_id=user_id,
-                    post_id=post_id
+            elif delete_value == "blog_comment":
+                res: str = str(query.get_blog_comment_author(
+                    user_id,
+                    post_id,
+                    delete_value
                 ))
+
+                if (session["user_id"] == user_id) or (session["user_id"] == res):
+                    query.delete_blog_comment(post_id, delete_value)
+
+                    return redirect(url_for(
+                        "blog",
+                        user_id=user_id,
+                        post_id=post_id
+                    ))
+            else:
+                return "Invalid value for field 'delete'", 400
 
         if request.method == "POST":
             return redirect(url_for("blog", user_id=user_id, post_id=post_id))
 
-    properties: list = profile.get_profile_properties(user_id)
+    properties: dict = profile.get_profile_properties(user_id)
     template: str = "blogpost.html"
     match properties["layout"]:
         case "twitter":
